Add ancient and lunar teleports
